ALLJOYN-1676 Copy the dbg file into the distribution for Windows builds
This change will makes it so SCons is now aware of the pdb files for
env.Program, and env.SharedLibrary builds, in both debug and release
mode.  Since these are part of the target output for the build they
will automatically be installed.

This change also has the good side effect that the -j scons option
is can now be used on a windows build since multiple files no longer
will share the same pdb file.

Change-Id: I005d40912111b41364129297d3395669d0ee1396