ALLJOYN-2143 Don't add any debug symbols to the release build
The debug symbols were added to both the release and debug build.  The release
build was not supposed to put any symbols in resulting binaries the debug
information was only supposed to be in the pdb file.  Which could be striped
out of the build. Tests showed it was not as easy to strip out as expected and
was causing the release build of code to be significantly larger than expected.

Change-Id: I3db5a75e64b2729e94216243e1e492b1163a22e4