ALLJOYN-2147 Add debug symbols to the obj files not separate pdb files
With a modern Microsoft compiler it is typical to use a pdb file i.e. the /Zi
or/ZI flags.  However, in SCons a pdb file is created for each .obj file.
To be able to use the debug information we would have to copy all of the
pdb files (one for each C++ file) into the dist. SCons documentation recommends
using the /Z7 option to solve this problem.  Since another more acceptable
solution has not yet been found we are going with the recommendation from the
SCons documentation.

Change-Id: Ib4aebc1da8134568b64216f31eae231509e4062a