ALLJOYN-2441 Improved the SCons Builder for Doxygen
The SCons Builder for Doxygen will now properly add dependencies and targets
when building HTML output.  Many of the tags from the configuration file are
not yet supported.  For a full list of the unprocessed tags  that could affect
the accuracy of the Builder see the final TODO comment in the
_doxygen_emitter function.

Change-Id: I62bda0108d11556822e063a5bb0d988d92712a5d