ALLJOYN-2459 Improved performance of whitespace checker on incremental builds
Do not run the whitespace checker on files that have already passed the
whitespace check and have not changed.  This builds a dictionary where
the key is the file name and the value is a sha1 hash.  If the sha1 hash
has not changed between runs then the file will not be checked for
whitespace compliance again.

This will create a file called whitespace.db in the folder SCons was run
from. This file contains the dictionary that will be used each check.

Only a minor performance decrease when doing a clean build. However,
a major performance increase for incremental builds.

Change-Id: I3511fddc9aa507444736e4fe57961a7b59e3c738