ALLJOYN-2441 Improved the SCons Doxygen Builder
The doxygen builder for SCons now understands how to build the list of
dependencies when building PDF output.

Fixed possible unclosed file pointer error.

Fixed issue that the 'recursive' tag was being read when the 'generate_html'
tag should have been read.

Script now uses the SCons env.Mkdir in place of the python os.mkdir.  This
should fix the issues seen when running this script from cygwin.

To find the PDF builder on windows SCons must be able to find the pdflatex tool
when the environment is created.  For this reason the script now finds all the
environment variables that need to be added to env['PATH'] before creating the
environment and adds them at creation time.

added DOXYGENFLAGS and DOXYGENCOM to the builder to these can be overridden if
needed.

Change-Id: Ie8064d85805b7cf7d39d200afea048199b064f6e