ALLJOYN-2441 Improved the SCons Doxygen Builder
Added a scanner to the Doxygen builder.  The emitter was doing work that the
scanner should have been doing.  The code is almost the same just split up.

The scanner is responsible for finding all of the files needed to build the
documentation while the emitter is used to find the input and output files
when the doxygen command is run. The emitter will not add files to the source
targets. While the scanner will.

Change-Id: Ie15ebd5513b50f68662b92652176b54de9452b2f